refactor: view 인터페이스화 및 AppConfig로 한 곳에서 관리하도록 변경
