+++ src/main/java/blackjack/Application.java
@@ -1,14 +1,14 @@
 package blackjack;
 
+import blackjack.config.AppConfig;
 import blackjack.controller.BlackjackController;
 import blackjack.util.Console;
-import blackjack.view.InputView;
-import blackjack.view.OutputView;
 
 public class Application {
 
     public static void main(String[] args) {
-        BlackjackController controller = new BlackjackController(new InputView(), new OutputView());
+        AppConfig config = new AppConfig();
+        BlackjackController controller = config.controller();
         try {
             controller.run();
         } finally {
