+++ src/main/java/blackjack/view/ConsoleOutputView.java
@@ -1,0 +1,66 @@
+package blackjack.view;
+
+import blackjack.dto.CardResult;
+import blackjack.dto.ProfitResult;
+import java.util.List;
+import java.util.StringJoiner;
+
+public class ConsoleOutputView implements OutputView {
+
+    public ConsoleOutputView() {
+    }
+
+    @Override
+    public void printInitialSettingsDoneMessage(String dealerName, List<String> playersName) {
+        System.out.println();
+        System.out.println(dealerName + "와 " + stringJoinWithComma(playersName) + "에게 2장을 나누었습니다.");
+    }
+
+    @Override
+    public void printCards(String playerName, List<String> cards) {
+        System.out.println(playerName + "카드: " + stringJoinWithComma(cards));
+    }
+
+    @Override
+    public void printCardsWithScore(CardResult cardResult) {
+        System.out.println(cardResult.participantName()
+                + "카드: " + stringJoinWithComma(cardResult.cardsName())
+                + " - 결과: " + cardResult.score());
+    }
+
+    @Override
+    public void printGetMoreCardsMessageForDealer(String dealerName) {
+        System.out.println();
+        System.out.println(dealerName + "는 16이하라 한장의 카드를 더 받았습니다.");
+    }
+
+    @Override
+    public void printResult(ProfitResult profitResult) {
+        System.out.println();
+        System.out.println("## 최종 수익");
+        printResultOfDealer(profitResult.calculateProfitOfDealer());
+        printResultOfPlayers(profitResult);
+    }
+
+    @Override
+    public void println() {
+        System.out.println();
+    }
+
+    private void printResultOfDealer(int profitOfDealer) {
+        System.out.println("딜러: " + profitOfDealer);
+    }
+
+    private void printResultOfPlayers(ProfitResult profitResult) {
+        profitResult.profitResult().forEach((playerName, profit) ->
+                System.out.println(playerName + ": " + profit));
+    }
+
+    private String stringJoinWithComma(List<String> strings) {
+        StringJoiner stringJoiner = new StringJoiner(",");
+        strings.forEach(stringJoiner::add);
+
+        return stringJoiner.toString();
+    }
+
+}
